+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the solarium library will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.0.1]
+### Fixed
+- Getting started documentation
+
+
 ## [5.0.0]
 ### Added
 - Component\Result\Facet\Bucket::getFacetSet()